Inertial Measurement Units (IMU) Data Package

This package is meant to serve as a data collection tool for IMU data. This data can be used as a means to assess and test methods designed to analyse IMU error signals (i.e. long and complex autocorrelated signals). An example method used for this kind of data is implemented in the GMWM R package which can also model the latent models that often characterize this data.

Details

Package: 	 gmwmdata

Type: 	 Package

Version: 	 3.0.0

Date: 	 2018-02-10

License: 	 CC BY-NC-SA 4.0

See also

cont.imu1, imu6, imar.gyro, ln200.accel, ln200.gyro, mtig, navchip and kvh

Examples

# Please see the datasets