IMU Data from a XSens MTi-G sensor

This data set contains gyroscope and accelerometer data from a XSens MTi-G sensor.

mtig

Format

An imu with 873,684 observations and 6 sensors. The sensors are defined as follows:

Gyro. X

Gyroscope X-Axis

Gyro. Y

Gyroscope Y-Axis

Gyro. Z

Gyroscope Z-Axis

Accel. X

Accelerometer X-Axis

Accel. Y

Accelerometer Y-Axis

Accel. Z

Accelerometer Z-Axis

Source

Geodetic Engineering Laboratory (TOPO), Swiss Federal Institute of Technology Lausanne (EPFL)

Details

This data set is equivalent to the imu6 data set. The only difference is that this data set acts as a way to showcase the results of the imu cast function whereas imu6 has yet to be cast using imu.

See also

mtig

Examples

# NOT RUN {
data(mtig)
mtig
# }